# address-service/src/main/java/com/nsc/address/model/address/Address.java
package com.nsc.address.model.address;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Address {
    private long id;
    private String city;
    private String district;
}